+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Changed
+
+- Update Bitmovin's native iOS SDK version to [`3.118.1`](https://developer.bitmovin.com/playback/docs/release-notes-ios#31181)
+
 ## [1.24.0] - 2026-07-31
 
 ### Added
@@ -11,7 +15,6 @@
 
 ### Changed
 
-- Update Bitmovin's native iOS SDK version to [`3.118.1`](https://developer.bitmovin.com/playback/docs/release-notes-ios#31181)
 - Cue layout no longer emits empty `layout` objects or uses `'auto'` for `line`, `position`, or `positionAlign`; automatic or unset placement is now represented by omitting these fields
 
 ## [1.23.0] - 2026-07-28