+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Update Bitmovin's native iOS SDK version to [`3.119.0`](https://developer.bitmovin.com/playback/docs/release-notes-ios#31190)
 
+### Fixed
+
+- iOS: Prevent startup crashes when Google Cast is not configured
+
 ## [1.24.0] - 2026-07-31
 
 ### Added